python/requirements.txt
Scrapy==1.8.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.8
prettytable==0.7.2
requests==2.22.0
urllib3==1.25.7
pymysql==0.9.3
psycopg2-binary==2.8.5
